+++ app/build.gradle
@@ -48,6 +48,8 @@ android {
 dependencies {
     implementation("androidx.core:core-ktx:1.13.1")
 
+    implementation("androidx.navigation:navigation-compose:2.7.7")
+
     implementation("androidx.lifecycle:lifecycle-runtime-ktx:2.8.4")
     implementation("androidx.activity:activity-compose:1.9.1")
     implementation("androidx.compose.ui:ui:$compose_version")
