+++ app/src/main/AndroidManifest.xml
@@ -14,7 +14,7 @@
         tools:targetApi="31">
 
         <activity
-            android:name=".MainActivity"
+            android:name=".presentation.MainActivity"
             android:exported="true"
             android:theme="@style/Theme.VkNewsClient">
             <intent-filter>
