+++ app/src/main/java/com/sumin/vknewsclient/domain/model/FeedPostModel.kt
@@ -1,4 +1,4 @@
-package com.sumin.vknewsclient.domain
+package com.sumin.vknewsclient.domain.model
 
 import com.sumin.vknewsclient.R
