+++ app/src/main/java/com/sumin/vknewsclient/presentation/VkNewsMainScreen.kt
@@ -1,0 +1,126 @@
+package com.sumin.vknewsclient.presentation
+
+import androidx.compose.foundation.clickable
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationItem
+import androidx.compose.material.Icon
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Scaffold
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.Favorite
+import androidx.compose.material.icons.outlined.Home
+import androidx.compose.material.icons.outlined.Person
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.saveable.rememberSaveable
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.vector.ImageVector
+import androidx.compose.ui.res.stringResource
+import androidx.navigation.compose.currentBackStackEntryAsState
+import androidx.navigation.compose.rememberNavController
+import com.sumin.vknewsclient.R
+import com.sumin.vknewsclient.presentation.navigation.AppNavGraph
+import com.sumin.vknewsclient.presentation.navigation.Screen
+
+
+@Composable
+fun MainScreen(viewModel: MainViewModel) {
+    val navHostController = rememberNavController()
+
+    Scaffold(
+        bottomBar = {
+            BottomNavigation {
+                val navBackStackEntry by navHostController.currentBackStackEntryAsState()
+                val currentRout = navBackStackEntry?.destination?.route
+                val items = listOf(
+                    NavigationItem.Home,
+                    NavigationItem.Favourite,
+                    NavigationItem.Profile
+                )
+
+                items.forEach { item ->
+                    BottomNavigationItem(
+                        selected = currentRout == item.screen.route,
+                        onClick = {
+                            navHostController.navigate(item.screen.route) {
+                                popUpTo(Screen.NewsFeed.route) {
+                                    saveState = true
+                                }
+                                launchSingleTop = true
+                                restoreState = true
+                            }
+                        },
+                        icon = {
+                            Icon(
+                                imageVector = item.icon,
+                                contentDescription = null
+                            )
+                        },
+                        label = {
+                            Text(text = stringResource(id = item.titleResId))
+                        },
+                        selectedContentColor = MaterialTheme.colors.onPrimary,
+                        unselectedContentColor = MaterialTheme.colors.onSecondary
+                    )
+                }
+            }
+        }
+    ) { paddingValues ->
+        AppNavGraph(
+            navHostController = navHostController,
+            homeScreenContent = {
+                ScreenHome(
+                    viewModel = viewModel,
+                    paddingValues = paddingValues
+                )
+            },
+            favouriteScreenContent = {
+                TextCounter(name = "Favourite")
+            },
+            profileScreenContent = {
+                TextCounter(name = "Profile")
+            })
+    }
+}
+
+
+sealed class NavigationItem(
+    val screen: Screen,
+    val titleResId: Int,
+    val icon: ImageVector
+) {
+    data object Home : NavigationItem(
+        screen = Screen.NewsFeed,
+        titleResId = R.string.navigation_item_main,
+        icon = Icons.Outlined.Home
+    )
+
+    data object Favourite : NavigationItem(
+        screen = Screen.Favourite,
+        titleResId = R.string.navigation_item_favourite,
+        icon = Icons.Outlined.Favorite
+    )
+
+    data object Profile : NavigationItem(
+        screen = Screen.Profile,
+        titleResId = R.string.navigation_item_profile,
+        icon = Icons.Outlined.Person
+    )
+}
+
+@Composable
+private fun TextCounter(name: String) {
+    var count by rememberSaveable {
+        mutableStateOf(0)
+    }
+
+    Text(
+        text = "$name Count: $count",
+        modifier = Modifier.clickable { count++ },
+        color = Color.Black
+    )
+}
