+++ app/src/main/java/com/sumin/vknewsclient/presentation/navigation/AppNavGraph.kt
@@ -1,0 +1,26 @@
+package com.sumin.vknewsclient.presentation.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+
+@Composable
+fun AppNavGraph(
+    navHostController: NavHostController,
+    homeScreenContent: @Composable () -> Unit,
+    favouriteScreenContent: @Composable () -> Unit,
+    profileScreenContent: @Composable () -> Unit
+) {
+    NavHost(navController = navHostController, startDestination = Screen.NewsFeed.route) {
+        composable(Screen.NewsFeed.route) {
+            homeScreenContent()
+        }
+        composable(Screen.Favourite.route) {
+            favouriteScreenContent()
+        }
+        composable(Screen.Profile.route) {
+            profileScreenContent()
+        }
+    }
+}
