+++ app/src/main/java/com/sumin/vknewsclient/presentation/navigation/Screen.kt
@@ -1,0 +1,13 @@
+package com.sumin.vknewsclient.presentation.navigation
+
+sealed class Screen(val route: String) {
+    data object NewsFeed : Screen(ROUTE_NEWS_FEED)
+    data object Favourite : Screen(ROUTE_FAVOURITE)
+    data object Profile : Screen(ROUTE_PROFILE)
+
+    private companion object {
+        const val ROUTE_NEWS_FEED = "news_feed"
+        const val ROUTE_FAVOURITE = "favourite"
+        const val ROUTE_PROFILE = "profile"
+    }
+}
