+++ app/src/main/java/com/sumin/vknewsclient/presentation/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.sumin.vknewsclient.ui.theme
+package com.sumin.vknewsclient.presentation.ui.theme
 
 import androidx.compose.ui.graphics.Color
