--- app/src/main/java/com/sumin/vknewsclient/ui/theme/VkNewsMainScreen.kt
@@ -1,166 +1,0 @@
-package com.sumin.vknewsclient.ui.theme
-
-import androidx.compose.foundation.ExperimentalFoundationApi
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material.BottomNavigation
-import androidx.compose.material.BottomNavigationItem
-import androidx.compose.material.DismissDirection
-import androidx.compose.material.ExperimentalMaterialApi
-import androidx.compose.material.FractionalThreshold
-import androidx.compose.material.Icon
-import androidx.compose.material.MaterialTheme
-import androidx.compose.material.Scaffold
-import androidx.compose.material.SwipeToDismiss
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.outlined.Favorite
-import androidx.compose.material.icons.outlined.Home
-import androidx.compose.material.icons.outlined.Person
-import androidx.compose.material.rememberDismissState
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.ui.Alignment
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.graphics.vector.ImageVector
-import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
-import com.sumin.vknewsclient.MainViewModel
-import com.sumin.vknewsclient.PostCard
-import com.sumin.vknewsclient.R
-
-
-@OptIn(ExperimentalMaterialApi::class, ExperimentalFoundationApi::class)
-@Composable
-fun MainScreen(viewModel: MainViewModel) {
-
-    Scaffold(
-        bottomBar = {
-            BottomNavigation {
-                val selectedItemPosition: MutableState<Int> = remember {
-                    mutableIntStateOf(0)
-                }
-
-                val items = listOf(
-                    NavigationItem.Home,
-                    NavigationItem.Favourite,
-                    NavigationItem.Profile
-                )
-
-                items.forEachIndexed { index, item ->
-                    BottomNavigationItem(
-                        selected = selectedItemPosition.value == index,
-                        onClick = { selectedItemPosition.value = index },
-                        icon = {
-                            Icon(
-                                imageVector = item.icon,
-                                contentDescription = null
-                            )
-                        },
-                        label = {
-                            Text(text = stringResource(id = item.titleResId))
-                        },
-                        selectedContentColor = MaterialTheme.colors.onPrimary,
-                        unselectedContentColor = MaterialTheme.colors.onSecondary
-                    )
-                }
-            }
-        }
-    ) {
-        val feedPostList = viewModel.feedPosts.collectAsState()
-
-        LazyColumn(
-            modifier = Modifier.padding(it),
-            contentPadding = PaddingValues(
-                top = 16.dp,
-                start = 8.dp,
-                end = 8.dp,
-                bottom = 16.dp
-            ),
-            verticalArrangement = Arrangement.spacedBy(8.dp)
-        ) {
-            items(feedPostList.value, key = { it.id }) { feedPostModel ->
-
-                val dismissState = rememberDismissState()
-                if (dismissState.isDismissed(DismissDirection.EndToStart)) {
-                    viewModel.deletePost(feedPostModel)
-                }
-                SwipeToDismiss(
-                    modifier = Modifier.animateItemPlacement(),
-                    state = dismissState,
-                    background = {
-                        Box(
-                            modifier = Modifier
-                                .padding(16.dp)
-                                .fillMaxSize()
-                                .background(Color.Red.copy(0.5f)),
-                            contentAlignment = Alignment.CenterEnd
-                        ) {
-                            Text(
-                                text = "Delete item",
-                                modifier = Modifier.padding(16.dp),
-                                fontSize = 24.sp,
-                                color = Color.White
-                            )
-                        }
-                    },
-                    directions = setOf(DismissDirection.EndToStart),
-                    dismissThresholds = {
-                        FractionalThreshold(0.5f)
-                    },
-                    dismissContent = {
-                        PostCard(
-                            feedPostModel = feedPostModel,
-                            onLikeClick = { statisticItem ->
-
-                                viewModel.updateCount(feedPostModel, statisticItem)
-                            },
-                            onCommentClick = { statisticItem ->
-                                viewModel.updateCount(feedPostModel, statisticItem)
-                            },
-                            onShareClick = { statisticItem ->
-                                viewModel.updateCount(feedPostModel, statisticItem)
-                            },
-                            onViewClick = { statisticItem ->
-                                viewModel.updateCount(feedPostModel, statisticItem)
-                            }
-                        )
-                    }
-                )
-            }
-        }
-    }
-}
-
-
-sealed class NavigationItem(
-    val titleResId: Int,
-    val icon: ImageVector
-) {
-
-    data object Home : NavigationItem(
-        titleResId = R.string.navigation_item_main,
-        icon = Icons.Outlined.Home
-    )
-
-    data object Favourite : NavigationItem(
-        titleResId = R.string.navigation_item_favourite,
-        icon = Icons.Outlined.Favorite
-    )
-
-    data object Profile : NavigationItem(
-        titleResId = R.string.navigation_item_profile,
-        icon = Icons.Outlined.Person
-    )
-}
